Creacion de modelos rol, usuario y mantenimiento de controlador producto

+++ src/com/appventas/modelo/Producto.java
@@ -1,4 +1,5 @@
 package com.appventas.modelo;
+
 import java.io.Serializable;
 import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.IntegerProperty;
@@ -25,7 +26,7 @@ import javax.persistence.Table;
 public class Producto implements Serializable {
     private IntegerProperty codigoProducto;
     private final StringProperty descripcion;
-    private final IntegerProperty precio;
+    private final DoubleProperty precio;
     private final DoubleProperty precioUnitario;
     private final IntegerProperty existencias;
     
@@ -34,14 +35,14 @@ public class Producto implements Serializable {
     public Producto(){
         this.codigoProducto = new SimpleIntegerProperty();
         this.descripcion = new SimpleStringProperty("");
-        this.precio = new SimpleIntegerProperty();
+        this.precio = new SimpleDoubleProperty();
         this.precioUnitario = new SimpleDoubleProperty();
         this.existencias = new SimpleIntegerProperty();
     }
-    public Producto(int codigoProducto, String descripcion, int precio, double precioUnitario, int existencias, Categoria categoria){
+    public Producto(int codigoProducto, String descripcion, double precio, double precioUnitario, int existencias, Categoria categoria){
         this.codigoProducto = new SimpleIntegerProperty(codigoProducto);
-         this.descripcion = new SimpleStringProperty (descripcion);
-        this.precio = new SimpleIntegerProperty(precio);
+        this.descripcion = new SimpleStringProperty (descripcion);
+        this.precio = new SimpleDoubleProperty(precio);
         this.precioUnitario = new SimpleDoubleProperty(precioUnitario);
         this.existencias = new SimpleIntegerProperty(existencias);
         this.categoria = categoria;
@@ -77,15 +78,15 @@ public class Producto implements Serializable {
     }
     
     @Column(name = "precio")
-    public int getPrecio(){
+    public double getPrecio(){
         return precio.get();
     }
     
-    public void setPrecio(int precio){
+    public void setPrecio(double precio){
         this.precio.set(precio);
     }
     
-    public IntegerProperty precio(){
+    public DoubleProperty precio(){
         return precio;
     }
